Assign fail timeouts to TinyPilot backends
If we don't explicitly define a fail_timeout for the backend services, nginx waits its default 10s before checking if a backend is available again. This means that if we do a simple restart of TinyPilot or uStreamer, nginx treats the services as down for 10 seconds.

This change reduces the timeout down to the lowest possible value, 1s, so nginx begins forwarding traffic as soon as possible.

https://github.com/mtlynch/tinypilot/issues/64