Verify HID module checksum exists before attempting to compare it
The HID module patch task can fail if the target system has just upgraded the kernel, as the path to the usb_f_hid.ko won't be correct until the following reboot. To avoid crashing the install with an 'undefined variable' error, we should first verify whether the variables we need are defined.

Fixes #164