Require collect-debug-logs to run as root
The main caller of the collect-debug-logs is the TinyPilot web app, which already executes collect-debug-logs as root.

The problem with sudo'ing mid-script is that it generates a lot of log spew about PAM sessions and executing as sudo. If we assume we're root throughout the script, it cuts down on log spew substantially.

Addresses #194 but does not completely fix it yet, as there is still some log spew from running the collect-debug-logs script itself.