Retrieve the uStreamer version from the binary instead of git
Versions of git with the 2022-04-12 security fix applied (https://github.blog/2022-04-12-git-security-vulnerability-announced/) will fail to check status if they're not run as the directory owner.

As a fix, we should just stop depending on git for the version number and use the output of ustreamer --version instead. This is also more flexible in the future if we decide to move from a source-compiled version to a precompiled Debian package for uStreamer.

Fixes https://github.com/tiny-pilot/tinypilot/issues/1140