Reorganize /opt/tinypilot-privileged
As part of transitioning privileged scripts from the Ansible role to the core tinypilot repo (see work-in-progress https://github.com/tiny-pilot/tinypilot/pull/1148), I'd like to adjust the file layout slightly to keep things tidy.

We have two types of executables in /opt/tinypilot-privileged: executables that the tinypilot user needs to execute with sudo privileges and executables that only root ever executes.

To make the distinction more obvious and make it easier to manage sudoers rules, I'd like to split the executables into two folders:

* /opt/tinypilot-privileged/scripts: For files that require root and the tinypilot user must execute them via sudo
* /opt/tinypilot-privileged/services: For files that require root but the tinypilot user never executes them

This PR moves the services file to the ./services subdirectory, which should not have any effect outside of the Ansible role.