Remove most tinypilot-privileged scripts from Ansible (#233)
This change gets rid of scripts and sudoers logic that allow the
`tinypilot` user to execute certain scripts as root. We are moving those
scripts and sudoers logic to the core TinyPilot repo:
https://github.com/tiny-pilot/tinypilot/pull/1148

We move `tinypilot_dir` and `tinypilot_privileged_dir` from `defaults`
to `vars` because they're no longer values that an Ansible playbook is
free to override, as the core TinyPilot repo assumes a directory of
`/opt/tinypilot` and a privileged directory of
`/opt/tinypilot-privileged`.